refactor(web): normalize route IA — flatten to top-level, /dashboard→/inboxes
The sidebar sections lived at two depths (Inboxes/Pending under
/dashboard/*, everything else top-level), and "back" from a message
detail went to the top-level list instead of the inbox. Normalize:

- /dashboard → /inboxes (the Inboxes list)
- /dashboard/agents/{messages,messages/view,settings} → /inboxes/{...}
  via a (view) route group so the list keeps its own layout (no agent
  chrome) while the per-inbox views share the agent layout.
- /dashboard/pending → /reviews; /dashboard/pending/review → /reviews/review
- /webhook-secrets → /webhooks
- breadcrumb is now context-aware: "Inboxes / {email}", where {email}
  links back to THAT inbox's message list — so detail/settings "back"
  returns to the inbox, not the top-level list.
- client-side redirect stubs at the old paths (/dashboard,
  /dashboard/pending, /webhook-secrets) since static export can't do
  server redirects; dropped the redundant /dashboard/agents index.

All entity routes keep ?email=/?id= query params (static export can't use
dynamic segments). Static export build green; tsc clean; 235 web tests.
