+++ web/src/app/(app)/get-started/_components/CustomAgentForm.tsx
@@ -41,7 +41,7 @@ export function CustomAgentForm({
         ...(name ? { name } : {}),
       });
       track("agent_creation_succeeded", { shared_or_custom: "custom", domain });
-      // Refresh the SWR `agents` cache so /dashboard shows the new
+      // Refresh the SWR `agents` cache so /inboxes shows the new
       // row immediately (otherwise keepPreviousData renders the
       // pre-create list until the next focus revalidation).
       await invalidateAgents();
