+++ web/src/app/(app)/inboxes/_components/AgentsEmptyState.tsx
