+++ web/src/app/(app)/inboxes/_components/ProtectionEditor.tsx
